# Bugfixes and speedups to _fuzz_search()