--- a/countrynames/data.yaml
+++ b/countrynames/data.yaml
@@ -18078,7 +18078,7 @@ NL:
   - Nedàlân
   - Niderland
   - Hà Lan
-NO:
+"NO":
   - Норвегия
   - Orílẹ́ède Nọọwii
   - নরওয়ে