--- a/test.py
+++ b/test.py
@@ -15,11 +15,11 @@ def test_to_code_3():
 
 
 def test_unicode():
-    assert to_code(u'Российская Федерация') == "RU"
+    assert to_code(u"Российская Федерация") == "RU"
 
 
 def test_fuzzy_matching():
-    assert to_code('Rossiyskaya Federatsiya', fuzzy=True) == "RU"
+    assert to_code("Rossiyskaya Federatsiya", fuzzy=True) == "RU"
     assert to_code("Falklands Islands", fuzzy=True) == "FK"
     assert to_code("TGermany", fuzzy=True) == "DE"
 
@@ -27,7 +27,7 @@ def test_fuzzy_matching():
 def test_non_standard_codes():
     assert to_code("European Union") == "EU"
     assert to_code_3("European Union") == "EUU"
-    assert to_code("Kosovo") == "XK"
+    assert to_code("Kosovo") == "XK", to_code("Kosovo")
     assert to_code_3("Kosovo") == "XKX"
 
 
@@ -36,8 +36,8 @@ def test_GB():
     assert to_code("Wales") == "GB-WLS"
     assert to_code("Northern Ireland") == "GB-NIR"
     assert to_code("Northern Ireland", fuzzy=True) == "GB-NIR"
-    assert to_code(
-        "United Kingdom of Great Britain and Northern Ireland") == "GB"
-    assert to_code(
-        "United Kingdom of Great Britain and Northern Ireland",
-        fuzzy=True) == "GB"
+    assert to_code("United Kingdom of Great Britain and Northern Ireland") == "GB"
+    assert (
+        to_code("United Kingdom of Great Britain and Northern Ireland", fuzzy=True)
+        == "GB"
+    )