There was a bug in `data.yml` that was preventing the data from being fully loaded. Apparently, YAML wants to turn the key literal "NO" into `False`. This was fixed be re-introducing quotes around the key.

Also, the `_fuzzy_search` method was optimized. It now runs at ~2x the speed simply by removing the second call to `Levenshtein.distance` and avoid using the `matches` list.

```
>>> # BEFORE:
>>> %timeit countrynames.to_code("Rossiyskaya Federatsiya", fuzzy=True)
44 ms ± 5.52 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
>>> # AFTER:
>>> %timeit countrynames.to_code("Rossiyskaya Federatsiya", fuzzy=True)
23.1 ms ± 631 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)
```